+++ lib/cargoDocTest.nix
@@ -25,5 +25,18 @@ mkCargoDerivation (
     pnameSuffix = "-doctest";
     buildPhaseCargoCommand = "";
     checkPhaseCargoCommand = "cargoWithProfile test --doc ${cargoExtraArgs} ${cargoTestExtraArgs}";
+
+    # Load bearing prepend here, in case the caller sets `preConfigure = "no newline";`
+    # we don't want to get a syntax error here (given how these are concatenated)
+    preConfigure = ''
+      if [[ -z "''${doCheck:-}" ]]; then
+        echo '#################################################'
+        echo '##                                             ##'
+        echo '## doCheck is not set, doc tests will not run! ##'
+        echo '##                                             ##'
+        echo '#################################################'
+      fi
+    ''
+    + (args.preConfigure or "");
   }
 )